Restructured to add a single Flee package
