+++ assets/behaviors/flee.behavior
