--- assets/behaviors/fleeInProximity.behavior
@@ -1,40 +1,0 @@
-{
-  "model": {
-    "nodeType": "engine:RepeatNode",
-    "nodeId": 0,
-    "node": {
-      "child": {
-        "nodeType": "engine:WrapperNode",
-        "nodeId": 1,
-        "node": {
-          "child": {
-            "nodeType": "engine:ParallelNode",
-            "nodeId": 2,
-            "node": {
-              "children": [
-                {
-                  "nodeType": "advancedBehaviors:checkFleeInProximityStopNode",
-                  "nodeId": 3,
-                  "node": {}
-                },
-                {
-                  "nodeType": "engine:RepeatNode",
-                  "nodeId": 0,
-                  "node": {
-                    "child": {
-                      "nodeType": "engine:LookupNode",
-                      "nodeId": 4,
-                      "node": {
-                        "tree": "Pathfinding:doRandomMove"
-                      }
-                    }
-                  }
-                }
-              ]
-            }
-          }
-        }
-      }
-    }
-  }
-}
