+++ src/main/java/org/terasology/advancedBehaviors/FleeOnHit/FleeOnHitComponent.java
@@ -25,10 +25,6 @@ import org.terasology.entitySystem.entity.EntityRef;
  * When it reaches a safe distance the instigator is set to null.
  */
 public class FleeOnHitComponent implements Component {
-    // Minimum distance from instigator after which the NPC will stop 'flee'ing
-    public float minDistance = 10f;
     // Speed factor by which flee speed increases
     public float speedMultiplier = 1.2f;
-    public EntityRef instigator;
-    public long timeWhenHit;
 }
